Fix OpenMP 2.0 unsigned iterator issues under Visual Studio

+++ src/plastimatch/bspline_mi.cxx
@@ -631,7 +631,7 @@ mi_hist_score_omp (BSPLINE_MI_Hist* mi_hist, int num_vox)
     double* m_hist = mi_hist->m_hist;
     double* j_hist = mi_hist->j_hist;
 
-    size_t f_bin, m_bin, j_bin;
+    int f_bin, m_bin, j_bin;
     double fnv = (double) num_vox;
     double score = 0;
     double hist_thresh = 0.001 / (mi_hist->moving.bins * mi_hist->fixed.bins);
@@ -1463,7 +1463,7 @@ bspline_score_f_mi (Bspline_parms *parms,
 {
     Bspline_score* ssd = &bst->ssd;
     BSPLINE_MI_Hist* mi_hist = &parms->mi_hist;
-    size_t pidx;
+    int pidx;
     float num_vox_f;
     Plm_timer timer;
 
@@ -1812,17 +1812,17 @@ bspline_score_e_mi (
     j_locks = (omp_lock_t*) malloc (mi_hist->fixed.bins * mi_hist->moving.bins * sizeof(omp_lock_t));
 
 #pragma omp parallel for
-    for (unsigned long i=0; i < mi_hist->fixed.bins; i++) {
+    for (long i=0; i < mi_hist->fixed.bins; i++) {
         omp_init_lock(&f_locks[i]);
     }
 
 #pragma omp parallel for
-    for (unsigned long i=0; i < mi_hist->moving.bins; i++) {
+    for (long i=0; i < mi_hist->moving.bins; i++) {
         omp_init_lock(&m_locks[i]);
     }
 
 #pragma omp parallel for
-    for (unsigned long i=0; i < mi_hist->fixed.bins * mi_hist->moving.bins; i++) {
+    for (long i=0; i < mi_hist->fixed.bins * mi_hist->moving.bins; i++) {
         omp_init_lock(&j_locks[i]);
     }
     /* ---------------------------------------------------- */
@@ -1899,7 +1899,7 @@ bspline_score_e_mi (
     }   // openmp
 
     /* Compute num_vox and find fullest fixed hist bin */
-    for (unsigned long i=0; i<mi_hist->fixed.bins; i++) {
+    for (long i=0; i<mi_hist->fixed.bins; i++) {
         if (f_hist[i] > f_hist[mi_hist->fixed.big_bin]) {
             mi_hist->fixed.big_bin = i;
         }
@@ -1907,7 +1907,7 @@ bspline_score_e_mi (
     }
 
     /* Fill in the missing histogram bin */
-    for (unsigned long i=0; i<mi_hist->moving.bins; i++) {
+    for (long i=0; i<mi_hist->moving.bins; i++) {
         mhis += m_hist[i];
     }
     m_hist[mi_hist->moving.big_bin] = (double)ssd->num_vox - mhis;
@@ -1915,7 +1915,7 @@ bspline_score_e_mi (
 
     /* Look for the biggest moving histogram bin */
 //    printf ("moving.big_bin [%i -> ", mi_hist->moving.big_bin);
-    for (unsigned long i=0; i<mi_hist->moving.bins; i++) {
+    for (long i=0; i<mi_hist->moving.bins; i++) {
         if (m_hist[i] > m_hist[mi_hist->moving.big_bin]) {
             mi_hist->moving.big_bin = i;
         }
@@ -1924,7 +1924,7 @@ bspline_score_e_mi (
 
 
     /* Fill in the missing jnt hist bin */
-    for (unsigned long j=0; j<mi_hist->fixed.bins; j++) {
+    for (long j=0; j<mi_hist->fixed.bins; j++) {
         for (unsigned long i=0; i<mi_hist->moving.bins; i++) {
             jhis += j_hist[j*mi_hist->moving.bins + i];
         }
@@ -1934,7 +1934,7 @@ bspline_score_e_mi (
     
     /* Look for the biggest joint histogram bin */
 //    printf ("joint.big_bin [%i -> ", mi_hist->joint.big_bin);
-    for (unsigned long j=0; j<mi_hist->fixed.bins; j++) {
+    for (long j=0; j<mi_hist->fixed.bins; j++) {
         for (unsigned long i=0; i<mi_hist->moving.bins; i++) {
             if (j_hist[j*mi_hist->moving.bins + i] > j_hist[mi_hist->joint.big_bin]) {
                 mi_hist->joint.big_bin = j*mi_hist->moving.bins + i;
@@ -1953,7 +1953,7 @@ bspline_score_e_mi (
     /* Display histrogram stats in debug mode */
     if (parms->debug) {
         double tmp;
-        unsigned long zz;
+        long zz;
         for (zz=0,tmp=0; zz < mi_hist->fixed.bins; zz++) {
             tmp += f_hist[zz];
         }
@@ -2071,17 +2071,17 @@ bspline_score_e_mi (
 
 
 #pragma omp parallel for
-    for (unsigned long i=0; i < mi_hist->fixed.bins; i++) {
+    for (long i=0; i < mi_hist->fixed.bins; i++) {
         omp_destroy_lock(&f_locks[i]);
     }
 
 #pragma omp parallel for
-    for (unsigned long i=0; i < mi_hist->moving.bins; i++) {
+    for (long i=0; i < mi_hist->moving.bins; i++) {
         omp_destroy_lock(&m_locks[i]);
     }
 
 #pragma omp parallel for
-    for (unsigned long i=0; i < mi_hist->fixed.bins * mi_hist->moving.bins; i++) {
+    for (long i=0; i < mi_hist->fixed.bins * mi_hist->moving.bins; i++) {
         omp_destroy_lock(&j_locks[i]);
     }
